[69] 채팅방리스트  프론트 구현
